Moved Settings to API folder/namespace
Because I don't think it should be in includes.
